补齐 M1 验收缺口、修正文档事实错误、修复符号分解复用缺陷
一、修一个真缺陷：运行时改 pin 掩码后没有重新做符号分解
   analyzed_ 一旦为真永不复位，而 assembleMatrix 只在**两端都自由**时写耦合块 -κI，
   所以给原本自由的顶点加 pin 会让 L 的稀疏结构变化。按 IGlobalSolver.h 自己记录的
   教训（符号结构与真实矩阵不一致 → solve 访问违例），此时必须重新 analyze。
   新增 Assembler::computeStructureStamp(拓扑 + 边表 + pin 掩码) 作为"结构戳"，
   Integrator 在结构变化时重新 analyze，数值分解行为不变。
   新增门禁 pinMaskChangeReanalyzesSparsityPattern：已注入旧行为验证它**会失败**。

二、补齐 M1 验收里一直没有测试的三项（spec 早已写好，代码从未实现）
   · globalSolveResidualIsMachinePrecision：‖Lx−b‖∞/‖b‖∞ < 1e-12，含反证
     （把解动 1e-6 时残差必须变大，保证判据有分辨力）；并验证复用分解解新右端。
   · iterationConvergesToFixedPointWithoutGravityOrDamping：12 点自由链，
     实测 463 次迭代收敛到 6.8e-13，且不动点使每条边回到 ℓ、总长 =(n−1)ℓ。
     **修正了 plan 里"k_max=200 且单调下降到 0"的错误规格**：本格式收敛是渐近的。
   · 收紧 iterationCountDoesNotChangeFixedPoint 的假断言：原容差 0.05，
     实测三个迭代数结果**逐位相同**（差 0），改为按机器精度断言。

三、文档事实校正（两个独立审计共报出约 40 处，此处改的是会造成误判的部分）
   plan.md
   · §1.2 单帧数据流图：b 与 L 的公式改为 M/h² 与 κ_c（原写成 m_v·x̂_v 与 M + ΣwAᵀA，
     既与代码不符，也与本文档 §2.2.1 自相矛盾）；速度更新补上 (1−k_d)
   · §2.1 顶点质量：补上"面积法 vs 边法差 2 倍"的口径说明（规则网格走边法，内部顶点
     2ρℓ²，而文档原公式给 ρℓ²）；帧预算自适应标注为未实现
   · §2.2.2 验收表改为与真实测试名一一对应，判据按实测口径重写；项 6 的投影方向
     由"沿 +ŷ"改为"沿 −ŷ（指向 pinned 端）"（原文与代码相反，会导致缺陷 A 复现）；
     项 7 由"无阻尼"改为实测用的 k_d=0.05；补上第 7/8/9 项
   · §2.4(4)：收益比公式与极限互为倒数，改为显式加速比并给出实测 9.0×
   · §2.4(6)：并行清单改为"现状 vs 计划"表 —— 当前只有局部步与散射真正并行
   · M1 验收 2/3/4/6 逐项标注 ✅/❌（2 与 6 的一半确实尚未实现）
   · M1.5：删除对第 7 项的误标；交叉引用错位（§2.2.3→§2.2.4 等 13 处）
   README.md
   · §3.4 chain 的"伸长相对误差"实为**总长**相对误差（两者差约 52 倍），改标签；
     删掉没有断言支持的"与解析 ε_k 吻合"
   code-map.md / Assembler.h / DistanceTerm.cpp / check.cpp
   · Assembler.h 头部约定块（"涉及 pinned 的约束既不进 L 也不进 b"、"右端 2κ_c d_c"）
     与代码和文档同时相反，是上一轮漏改的 —— 照着改会把右端乘 2 或整条跳过弹簧约束
   · DistanceTerm.cpp 两处"配错符号会跑到 ℓ−mg/κ"写反了：ℓ−mg/κ 正是**正确**平衡，
     符号反了应跑到 ℓ+mg/κ
   · check.cpp 头部注释写"平衡 = ℓ+mg/κ"，与它自己第 3 项的判据相反
   · Mesh.cpp 质量注释"边法与三角形法一致"不成立（差 2 倍）
   · one_step_trace.cpp 里把 ℓ+mg/κ 标为"候选平衡"，改为标注已废弃

四、构建卫生
   · 消掉全部编译警告（未引用形参、%zu 格式、未引用变量、getenv 弃用）
   · test_harness.h：checkCount() 是 int 而格式用 %zu，显式转换

回归：pd_check 8/8、test_primitives 68 断言、test_spring_vertical 155 断言、
pd_trans / pd_chain / pd_diraudit 全部 exit 0；pd_bench 数值分解仍为 1 次。

+++ tests/primitives/test_harness.h
@@ -49,8 +49,7 @@ inline void reportFailure(const char* file, int line, const std::string& what) {
 }
 
 inline int runAll(const char* suiteName) {
-  std::printf("=== %s: %zu 个测试 ===\n", suiteName, registry().size());
-  int failedTests = 0;
+  std::printf("=== %s: %zu 个测试 ===\n", suiteName, registry().size());  int failedTests = 0;
   for (auto& t : registry()) {
     const int before = failureCount();
     std::printf("  [run ] %s\n", t.name.c_str());
@@ -62,8 +61,9 @@ inline int runAll(const char* suiteName) {
       std::printf("  [ok  ] %s\n", t.name.c_str());
     }
   }
-  std::printf("--- %s: %zu 个断言, %d 个失败测试, %d 个失败断言 ---\n", suiteName, checkCount(),
-              failedTests, failureCount());
+  // checkCount() 返回 int，而 %zu 期待 size_t —— 显式转换，免得换编译器/换警告等级时炸出格式警告。
+  std::printf("--- %s: %zu 个断言, %d 个失败测试, %d 个失败断言 ---\n", suiteName,
+              static_cast<std::size_t>(checkCount()), failedTests, failureCount());
   return failedTests == 0 ? 0 : 1;
 }
